chore: use mise for contracts deps (#1635)
* chore: add fmt-check contract to makefile

* fix(make): fmt-check

Previously the gofmt -l command was running through every single go file in subdirs, including contract submodules. It was erroring on a non-formatted file from eigenlayer-middleware I believe, or some other contract dep libs.

This didn't happen in CI because the ci that runs this doesn't checkout recursively the repo. But it happened locally and was annoying. So this fixes it.

* chore: add contract tooling deps to mise

* ci(contracts): turn back on bindings diff check

I believe this was turned off because bindings had small diffs between local and CI. My hunch is that this was because of using different solidity versions locally and in CI. Using mise to pin down the forge version should hopefully fix this... so readding this check back.

Also using mise-install to install all other dependencies that were installed with github actions (abigen, node, yarn, etc)

* chore: make compile-contracts

* ci(contracts): print forge/abigen versions to debug

* ci(contracts): turning back binding diff check off (I failed..)

No idea why this doesnt work. I give up.

* ci(contracts): only run push on master

Was running CI twice for every PR because it already runs on `pull_request` events.

* chore(mise): bump node dep 16->18

integration test was failing because some go dependency for subgraph requires a node engine >= 18.0.0

* chore(mise): bump node 18->20

another inabox subgraph error hit: 2025/06/03 12:22:30 exit status 1: error @graphprotocol/graph-cli@0.96.0: The engine "node" is incompatible with this module. Expected version ">=20.18.1". Got "18.20.8"

* chore: remove bytecode from bindings so they only contain abi (#1636)

* chore: change contracts compile.sh to generate bindings with only abi

We remove the bytecode from bindings. This was causing all sorts of CI issues (see https://github.com/Layr-Labs/eigenda/pull/1635).

I believe we weren't deploying any contracts from golang code, so this should be fine. Anyone who does want to do this can create their own "deployable-bindings" (we could also create these and check them in to the repo if needed...)

* chore: make compile-contracts

Regenerate bindings to remove bytecode, so they only contain abi now.

* ci(contracts): turn on bindings diff check

This was turned on because ci was choking on some bytecode changes that we don't understand. Now that are bindings only contain abi, this test should be turned on again.

* fix: contracts/compile.sh script