fix(dataapi): fix time query_param serialization/formatting bugs (#1672)
fix(dataapi): fix time queryparam serialization/formatting bugs

Some weird handwritten format was being used before, which was not RFC3399Nano... and was causing bugs in the test (passing in CI only because ci is running in north america). I noticed the bug when trying to run the tests locally from Indonesia.