chore: bump to go1.24 (#1669)
* chore: bump to go1.24 and toolchain go1.24.4

See comments in go.mod for why we are doing this

* chore: bump dockerfiles to use go1.24

Don't think this is strictly necessary because I think the `go` binary inside the containers is already future compatible (https://go.dev/blog/toolchain), but some dockerfile was failing and didn't feel like debugging so just bumped their versions.

* chore: bump mise go version to 1.24 + document

Document why this isn't strictly necessary in the mise file itself.

* ci: remove unneeded input from test-coverage/action.yml

* fix(mise): abigen version compatible with go1.24

* fix: bump golangci-lint version so we can install with go1.24

* fix: lint issues from new linters

* fix: err condition in ReverseBitOrder

* fix(benchmark): ineffective break -> return

* Fix bug in reverse bits logic

* test: fix test Missing SvcManagerAddr error

* fix: encoderClient bug introduced in previous commit

refactored a line but forgot the defer, so the connection was getting closed immediately, causing the client encodeBlob request to fail before being sent.

* style: wrap errors for easier debugging

---------

Co-authored-by: Cody Littley <cody@eigenlabs.org>