fix: load generator improvements (#1722)
* fix: defensive NewCertBuilder

Wasn't sanitizing its passed inputs before, which caused a very hard to debug error when deploying to preprod where the RegistryCoordinatorAddr was not populated on the load generator.

* chore: load_generator makefile improvements

Only rebuild when necessary, and make build the default command, as opposed to clean

* fix: loadgen metrics nil ptr bug

load generator was crashing when metrics were set to false. Believe this is a mistake in prometheus' promauto API. created an issue for this upstream (see comment in code), and added a hacky fix in our repo in the meantime.

* test: NewTestClient uses RegistryCoordinatorAddr from ethReader

Instead of using one that was read from the config, which was bug prone (and did lead to a bug)

* chore: CertBuilder info logs eth-call arguments

This is needed to debug this call. We had one failing on preprod which was failing with a "execution reverted". We should probably add eth-calls like this before every eth-call site. Unsure if we want info or debug yet...

* test: remove RegistryCoordinator addr from TestClientConfig

No longer needed as we read it from the ServiceManager. This is less bug prone.

* chore: make fmt

* fix: relative paths in test/v2/config files

Before relative paths wrt test files were used in the config file (for SRSPaths), which only worked for the tests, but not for running the load-generator.

I made the paths relative to the config file itself, so now they work for both tests and load-generators. I think it also makes understanding the config files easier (otherwise there's some mysterious relative path which doesn't make sense without knowing that the paths are meant to be interpreted wrt where the tests are located).

* style(cert_builder): only log eth-call params on error

This way we only log when actually needed (to debug), and don't clutter our happy-path logs for no reason.

* chore: fix test/v2/Makefile test command