chore(proxy): consolidate dotenv examples files into single one (#1800)
* chore(proxy): simplify .env example files to a single one

Having multiple example files is a source of bug and confusion. its also hard to maintain. Consolidating into a single .env.example file. People should figure out how to populate it for their own network anyways.

Also removed all the config values that have sane defaults, and only left the ones that are mandatory to set. This removes a lot of cognitive load on our users, and will be a helpful simplification I believe.

* fix(proxy): cache/fallback cli ingestion bug

Very minor bug that was being misread. belive its a urfave bug but havent spent more time investigating.

* fix: build

* chore(proxy): make gen-static-help-output

* chore: proxy turn on metrics server by default (#1801)

* chore(proxy): enable metrics server by default

This simplifies the .env.example config file, and makes one less configuration to search for for new users.

* chore: make gen output

* docs(proxy): fix readme numbering

* docs(.env.example): EIGENDA_PROXY_STORAGE_BACKENDS_TO_ENABLE comment

* docs(.env.example): EIGENDA_PROXY_STORAGE_DISPERSAL_BACKEND comment

* docs(.env.example): EIGENDA_PROXY_EIGENDA_V2_MAX_BLOB_LENGTH comment

* chore(.env.example): switch mainnet -> sepolia default values