test: add payload-to-blob conversion test_example (#1833)
* test: add payload-to-blob conversion test_example

Hard to understand and follow the conversion process that a payload follows when being converted to a blob. This is a first step in adding tests and examples to show this. This specific example only shows the route when interpreting the encodedPayload as evaluations and taking the IFFT to get a blob (polynomial coefficients).

* style: payload head -> EncodedPayloadHeader