feat(metrics): Add V2 accountant metrics (#1864)
* feat(metrics): Add v2 accountant metrics

Adds v2 accountant metrics, starting with cumulative payment

* Return `noop` instead of `nil`

Returns a `noopAccountantMetricer` if the `registry` passed to
`NewAccountantMetrics` is `nil`

* refactor: create accountant before constructing disperser client

Refactors accountant construction so it takes place before the
disperser client is constructed.

Updates PopulateAccountant to return an error is accountant is nil

Removes metric setters

* lint: fix lint issues

* Update gwei convertion in `RecordCumulativePayment`

* Add `NewUnpopulatedAccountant` function

Adds a `NewUnpopulatedAccount` function that simply calls the
`NewAccountant` function with default arguments.