feat(proxy)!: remove support for redis cache (#1892)
feat!: remove support for redis cache

We asked and no customer is using it. Removing this will allow for refactoring of secondary storage backends, and other generic cleanup of the code. Trying to keep proxy minimal with only required feature support. This minimizes code that needs to be audited, and chance of bugs creeping in.