fix(cert-verifier): add try/catch to never revert and classify all checkDACert calls as valid/invalid/bug (#1938)
* refactor: cert verifier lib reverts, certVerifier try/catches

This is meant to keep the library cleaner, and using modern custom error practices. We only need the outside CertVerifier to not revert, which we achieve with a try/catch.

* try/catch abi.decode

by moving it to CertVerifier and making it external

* docs: document EigenDACertVerifier contract

* bump CertVerifier semver

* certVerifier constructor reverts on InvalidQuorumNumbersRequired

* certVerifierLib: robustify checkSecurityParams

* simplify and document quorum subsets

* document external block.number > RBN requirement

* feat: add bug status code instead of rethrowing

After discussing with ethen realized this will have better behavior as it will make it impossible for a malicious batch poster to stall the pipeline.

We can also match BUG status codes to failover signals on the batcher write path.

* remove unneeded comment

* test: fix unit test

* style: make fmt

* regenerate bindings

* add comment to get rid of UNUSED statuscodes when bump to 4.0.0

* move contract docs to checkDACert

* easier to understand INVALID_CERT comment

* fix offchain to understand new status codes

* fix inabox certverification error checks

* rename BUG -> INTERNAL_ERROR

* rename decodeCert -> _decodeCert

Also mention that it isnt meant to be part of the public API.

* better explanation for OOG error TODO

* style: fix line length lint errors

* gen bindings