fix(bloblength): use uint32 instead of uint (#2053)
* fix(blobLength): use uint32 instead of uint

We've been using uint/int in a bunch of data types which are fundamental to our protocol, and influence hashing and possibly other things. We shouldn't.

bloblength is once instance where we use uint32 onchain, uint offchain (and actually uint64 in the encoder proto!).

This PR changes the main offchain bloblength type from uint to uint32, and propagates the changes everywhere.

I decided not to change all the v1 chunk encoding/assignment logic which still uses uint, and so just convert uint32->uint when calling those functions. This is fine because golang uint type is always at least uint32.

* fix lint

* style: fix proto lint issue

* test: fix encoding_streamer_test

* fix lint