fix(inabox): network issues (#2069)
* chore(inabox): make deploy-all creates new-anvil first

This is needed because running deploy-all first, then killing chain, and restarting it causes a bug because the deploy code reads a config which tells it the contracts are already deployed, and it skips deploying (because it doesnt know its a new chain).

* fix(inabox): create chain network

There was actually a few errors that we ran into. graph couldnt connect to chain. added better error handling as well in deploy-all.

Most importantly creating a network instance for anvil and graph to run in. test-containers doesn't allow naming this network... so we end up with a lot of garbage UUID docker networks... these will need to be pruned individually (or by `docker system prune -A`).